Fix slowdown by moving overlays less often.
Now a new function `hl-paren-initiate-highlight` is in `post-command-hook`
instead of `hl-paren-highlight` itself.  The new function uses a timer to skip
calls to `hl-paren-highlight` in case those come faster than about one in a
quarter of a second.  That easily happens when scrolling by pressing and holding
`C-n`.

Fixes issue #8.

In contrast to the subject of the issue, I don't use an idle timer but a normal
one to ensure that the hl-paren updates are immediately visible from a user's
point of view.